add gatekeeper to cli
